- added game states - pause/resume - update mixin to unify updates

+++ lib/levels/level_parameters.dart
@@ -24,7 +24,7 @@ class TrashObjective
 {
   TrashType trashType;
   int goal;
-  int timeLimit;
+  double timeLimit;
 
   TrashObjective({
     required this.trashType,
